test(mfa): restructure test foundation by domain guarantee (screaming architecture)
The layout named files after the xstate/jest mechanism (machine / coverage /
contract / model), so the folder screamed 'xstate test kit' rather than 'MFA
modal flow' and gave no clue what the suite guarantees. Reorganize it around the
two domain questions it answers.

Specs (tests/unit/MultifactorAuthentication) split into two domain folders:
- modalFlow/ - what the machine does: scenarioOpensModalTest, modalLifecycleTest,
  everyStateReachableTest. The old coverage file mixed reachability and the
  lifecycle lap; split into one guarantee per file.
- viewMatchesMachine/ - every reader agrees with the machine, in three layers of
  realism: consumerStateTest (snapshotToState), reactBindingTest (@xstate/react),
  realModalTest (real provider + gestures + DOM).

Harness (tests/utils/mfa, the repo's helper home since tests/unit/** collects
every file as a suite) renamed for intent: flowFixtures, machineUnderTest,
flowPaths, reachableStates, realUi/{renderModal,userGestures,mocks,jestMocks}.

Add a README with the current state chart, the two guarantee families, a harness
map, and a 'where to add a new screen/slice' guide.

Pure test-only refactor: no production code touched, exported symbols unchanged,
and no assertion changed beyond mechanically splitting the coverage file. All 12
tests pass; prettier and typecheck clean.
