+++ tests/unit/MultifactorAuthentication/viewMatchesMachine/consumerStateTest.ts
@@ -1,11 +1,11 @@
 import {snapshotToState} from '@components/MultifactorAuthentication/machine';
 import CONST from '@src/CONST';
-import createMfaTestActor from '../../utils/mfa/createMfaTestActor';
-import createInitEvent, {MFA_TEST_SCENARIO_NAME} from '../../utils/mfa/mfaTestFixtures';
+import createInitEvent, {MFA_TEST_SCENARIO_NAME} from '../../../utils/mfa/flowFixtures';
+import createMfaTestActor from '../../../utils/mfa/machineUnderTest';
 
 const MFA_STATE = CONST.MULTIFACTOR_AUTHENTICATION.MFA_STATE;
 
-describe('snapshotToState', () => {
+describe('consumers read the modal state from the machine', () => {
     it('derives modalState from the machine state and flattens the context', () => {
         const {actor} = createMfaTestActor();
         expect(snapshotToState(actor.getSnapshot()).modalState).toBe(MFA_STATE.CLOSED);
