+++ tests/unit/MultifactorAuthentication/viewMatchesMachine/realModalTest.tsx
@@ -1,21 +1,21 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-return -- jest.mock factories delegate to require()'d helpers, which resolve as `any`. */
 import {resetMfaNavigation} from '@components/MultifactorAuthentication/mfaNavigation';
 import CONST from '@src/CONST';
-import createMfaTestModel from '../../utils/mfa/createMfaTestModel';
-import mfaEventExecutors from '../../utils/mfa/ui/eventExecutors';
-import {flushMfaUi, isModalOverlayMounted, isOutcomeScreenVisible, renderMfaUi} from '../../utils/mfa/ui/harness';
-import {resetMfaUiMocks} from '../../utils/mfa/ui/mocks';
+import createMfaTestModel from '../../../utils/mfa/flowPaths';
+import {resetMfaUiMocks} from '../../../utils/mfa/realUi/mocks';
+import {flushMfaUi, isModalOverlayMounted, isOutcomeScreenVisible, renderMfaUi} from '../../../utils/mfa/realUi/renderModal';
+import mfaEventExecutors from '../../../utils/mfa/realUi/userGestures';
 
 // Wide layout so the navigator renders the backdrop overlay the assertions use as the mounted marker.
 jest.mock('@hooks/useResponsiveLayout');
 // Dev-only Stately inspector wiring -> the plain @xstate/react adapter the provider needs.
-jest.mock('@hooks/useInspectedMachine', () => require('../../utils/mfa/ui/jestMocks').inspectedMachineMock());
+jest.mock('@hooks/useInspectedMachine', () => require('../../../utils/mfa/realUi/jestMocks').inspectedMachineMock());
 // Native / WebAuthn biometrics are out of scope for the modal-lifecycle contract.
-jest.mock('@components/MultifactorAuthentication/biometrics/useBiometrics', () => require('../../utils/mfa/ui/jestMocks').biometricsHookMock());
+jest.mock('@components/MultifactorAuthentication/biometrics/useBiometrics', () => require('../../../utils/mfa/realUi/jestMocks').biometricsHookMock());
 // Browser/Android back-history wiring is a separate concern from the machine <-> UI contract.
-jest.mock('@components/MultifactorAuthentication/useSyncMfaModalNavigatorWithHistory', () => require('../../utils/mfa/ui/jestMocks').syncHistoryMock());
+jest.mock('@components/MultifactorAuthentication/useSyncMfaModalNavigatorWithHistory', () => require('../../../utils/mfa/realUi/jestMocks').syncHistoryMock());
 // Navigation automock leaves methods undefined; supply the three the flow needs and no-op the rest.
-jest.mock('@libs/Navigation/Navigation', () => require('../../utils/mfa/ui/jestMocks').navigationMock());
+jest.mock('@libs/Navigation/Navigation', () => require('../../../utils/mfa/realUi/jestMocks').navigationMock());
 
 const MFA_STATE = CONST.MULTIFACTOR_AUTHENTICATION.MFA_STATE;
 
@@ -42,7 +42,7 @@ const testConfig = {
     },
 };
 
-describe('mfaMachine driven through the real MFA UI', () => {
+describe('the real MFA modal follows the machine', () => {
     beforeEach(() => {
         resetMfaUiMocks();
         resetMfaNavigation();
